Move "Setting" and "SettingSwitch" components to GameplaySettings modal components folder
