+++ frontend/src/features/modals/GameplaySettings/GameplaySettings.tsx
@@ -1,8 +1,8 @@
-import useGameplaySettings from "../../hooks/useGameplaySettings.js";
-import Setting from "../../components/global/settings/Setting.js";
+import useGameplaySettings from "../../../hooks/useGameplaySettings.js";
+import Setting from "./components/Setting.js";
 
 import "../../../styles/modals/gameplay-settings.scss";
-import { StateSetterFunction } from "../../types/general.js";
+import { StateSetterFunction } from "../../../types/general.js";
 
 type GameplaySettingsModalProps = {
 	onClose: () => void;
