+++ frontend/src/features/modals/GameplaySettings/components/SettingsSwitch.tsx
